Initial upload
This project can be loaded directly into Netbeans. If using another IDE, the netbeans folder can be ignored. All of the code and the starting CSS file is contained in SRC. The FRCTeams csv file is used for question generating. Programmers are encouraged to replace this with their own question generation code.